update lambda function to use FILE_SOURCE environment variable to drive varying execution

+++ lambda_function.py
@@ -15,7 +15,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from urllib.parse import urlparse, urlunparse
 
-FILE_SOURCE = os.environ.get("FILE_SOURCE", "local") # alternative is google_drive
+FILE_SOURCE = os.environ.get("FILE_SOURCE", "google_drive")  # alternative is local
 PRODUCTION = os.environ.get("PRODUCTION", "0") == "1"
 TEMP_FOLDER = "/tmp" if PRODUCTION else os.path.dirname(__file__)
 WEBDRIVER_TIMEOUT_SECONDS = int(os.environ.get("WEBDRIVER_TIMEOUT_SECONDS", "15"))
@@ -420,29 +420,44 @@ class BizBuySellAutomator:
         return downloadable_link
 
     def automate_single_user_session(
-        self, username: str, password: str, csv_link: str
+        self,
+        username: str,
+        password: str,
+        csv_link: str,
     ) -> None:
         """Automates batch upload session for a single user
         Args:
         username (str) - user's username or email address
         password (str) - user's password
-        csv_link (str) - google drive csv link for this user to batch upload
+        csv_link (str) - google drive csv link for this user to batch upload (FILE_SOURCE=google_drive) OR local path (FILE_SOURCE=local)
 
         Returns: None
         """
         self.info(f"Automating user session for {username}")
 
         self.login(username=username, password=password)
-        # Download the CSV for this user with the URL from the Lambda environment
-        csv_file_path = self.google_drive_downloader.download_file_from_google_drive(
-            shared_link=csv_link
-        )
 
-        # Automate the upload of that CSV with the current user's web app session
+        if FILE_SOURCE == "google_drive":
+            # file not already on file system
+            # Download the CSV for this user with the URL from the Lambda environment
+            csv_file_path = (
+                self.google_drive_downloader.download_file_from_google_drive(
+                    shared_link=csv_link
+                )
+            )
+        elif FILE_SOURCE == "local":
+            # Already stored locally. Ensure path exists before using.
+            assert os.path.exists(csv_link)
+            csv_file_path = csv_link
+
+        # Automate the upload of that CSV on local path with
+        # the current user's web app session
         self.automate_upload(csv_file_path=csv_file_path)
 
-        # Now remove the temporary downloaded file
-        os.remove(csv_file_path)
+        # IF the file was downloaded from Google Drive,
+        # remove the temporary downloaded file
+        if FILE_SOURCE == "google_drive":
+            os.remove(csv_file_path)
 
         self.logout()
 
@@ -487,11 +502,10 @@ class BizBuySellAutomator:
 
         self.driver.close()
         self.driver.quit()
- 
 
 
-class Driver: 
-    def __init__(self, verbose: bool = False): 
+class Driver:
+    def __init__(self, verbose: bool = False):
         self.verbose = verbose
         self.setup_logging()
 
@@ -521,20 +535,20 @@ class Driver:
 
     def error(self, msg) -> None:
         self.logger.error(msg, extra={"prefix": self.name})
-        
-    def run_local(self) -> None: 
-        """ Method to run the automation on a local server without AWS lambda.
-        Uses environment variables instead of lambda event to drive execution """
+
+    def run_local(self) -> None:
+        """Method to run the automation on a local server without AWS lambda.
+        Uses environment variables instead of lambda event to drive execution"""
         self.info("Running local execution with values from environment variables")
-        mode = os.environ.get("MODE", "single_user") # alt = multi_user
-        if mode == "single_user": 
+        mode = os.environ.get("MODE", "single_user")  # alt = multi_user
+        if mode == "single_user":
             try:
                 assert all(
                     x in os.environ
                     for x in [
-                        "single_user_password",
-                        "single_user_username",
-                        "single_user_csv",
+                        "SINGLE_USER_USERNAME",
+                        "SINGLE_USER_PASSWORD",
+                        "SINGLE_USER_CSV",
                     ]
                 )
             except AssertionError as e:
@@ -550,6 +564,41 @@ class Driver:
                         )
                     },
                 }
+            logger.info("Creating automator with MODE=single_user", extra=extra)
+            try:
+                automator = BizBuySellAutomator(verbose=verbose)
+                automator.init_driver()
+                automator.automate_single_user_session(
+                    username=os.environ.get("SINGLE_USER_USERNAME", ""),
+                    password=os.environ.get("SINGLE_USER_PASSWORD", ""),
+                    csv_link=os.environ.get("SINGLE_USER_CSV", ""),
+                )
+                automator.quit()
+                return {
+                    "statusCode": 200,
+                    "headers": {"Content-Type": "application/json"},
+                    "body": {
+                        "success": (
+                            f'batch upload of {os.environ.get("SINGLE_USER_CSV")}'
+                            f' complete for single_user {os.environ.get("SINGLE_USER_USERNAME")}'
+                        )
+                    },
+                }
+            except TimeoutException as e:
+                logger.error(traceback.format_exc(), extra=extra)
+                return {
+                    "statusCode": 500,
+                    "headers": {"Content-Type": "application/json"},
+                    "body": {"error": traceback.format_exc()},
+                }
+            except Exception as e:
+                logger.error(traceback.format_exc(), extra=extra)
+                return {
+                    "statusCode": 500,
+                    "headers": {"Content-Type": "application/json"},
+                    "body": {"error": traceback.format_exc()},
+                }
+
         elif mode == "multi_user":
             try:
                 assert "MULTI_USER_CSV" in os.environ
@@ -565,21 +614,26 @@ class Driver:
                             "username,password,csv_link as columns"
                         )
                     },
-                } 
+                }
             try:
                 logger.info("Creating automator with mode=multi_user", extra=extra)
                 automator = BizBuySellAutomator(verbose=verbose)
                 automator.init_driver()
-                # Download the CSV for multi-user execution
-                # should be formatted as username,password,csv_link where
-                # csv_link is the batch upload file for that user
-                multi_user_csv_path = (
-                    automator.google_drive_downloader.download_file_from_google_drive(
-                        shared_link=event["multi_user_csv"],
+                if FILE_SOURCE == "google_drive":
+                    # Download the CSV for multi-user execution
+                    # should be formatted as username,password,csv_link where
+                    # csv_link is the batch upload file for that user
+                    multi_user_csv_path = automator.google_drive_downloader.download_file_from_google_drive(
+                        shared_link=os.environ.get("MULTI_USER_CSV", ""),
                         temporary_filename="multi-user-tmp.csv",
                     )
+                elif FILE_SOURCE == "local":
+                    # use the local FS path to the file; csv_link column should also specify local FS paths
+                    # for each user
+                    multi_user_csv_path = os.environ.get("MULTI_USER_CSV", "")
+                automator.automate_multiple_user_sessions(
+                    csv_file_path=multi_user_csv_path
                 )
-                automator.automate_multiple_user_sessions(csv_file_path=multi_user_csv_path)
                 automator.quit()
                 return {
                     "statusCode": 200,
@@ -601,8 +655,8 @@ class Driver:
                     "body": {"error": traceback.format_exc()},
                 }
 
-    def run_lambda(self, event, context) -> None: 
-        """ Run automation with AWS lambda using event to drive execution """
+    def run_lambda(self, event, context) -> None:
+        """Run automation with AWS lambda using event to drive execution"""
         self.info("Running AWS Lambda execution with values from event")
         if "mode" not in event or event["mode"] == "single_user":
             try:
@@ -677,7 +731,7 @@ class Driver:
                             "username,password,csv_link as columns"
                         )
                     },
-                } 
+                }
             try:
                 logger.info("Creating automator with mode=multi_user", extra=extra)
                 automator = BizBuySellAutomator(verbose=verbose)
@@ -691,7 +745,9 @@ class Driver:
                         temporary_filename="multi-user-tmp.csv",
                     )
                 )
-                automator.automate_multiple_user_sessions(csv_file_path=multi_user_csv_path)
+                automator.automate_multiple_user_sessions(
+                    csv_file_path=multi_user_csv_path
+                )
                 automator.quit()
                 return {
                     "statusCode": 200,
@@ -714,21 +770,16 @@ class Driver:
                 }
 
 
-
-
-
 def lambda_handler(event, context):
     """Primary handler function for AWS Lambda to execute"""
     verbose = False
     if "verbose" in event:
         verbose = event["verbose"]
-
     driver = Driver(verbose=verbose)
     driver.run_lambda(event, context)
-      
- 
-     
+
+
 if __name__ == "__main__":
-    verbose = os.environ.get("VERBOSE", "0") == "1" 
+    verbose = os.environ.get("VERBOSE", "0") == "1"
     driver = Driver(verbose=verbose)
     driver.run_local()
